+++ src/main/java/org/realityforge/guiceyloops/server/glassfish/GlassFishContainer.java
@@ -34,7 +34,7 @@ public class GlassFishContainer
   public GlassFishContainer( final int port )
     throws Exception
   {
-    this( port, GlassFishVersion.V_4_1 );
+    this( port, GlassFishContainerUtil.getEmbeddedGlassFishClasspath() );
   }
 
   public GlassFishContainer( final int port, @Nonnull final GlassFishVersion version )